requirements.txt
pandas==2.1.4
numpy==1.26.4
faker==24.4.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
python-dotenv==1.0.1
pyyaml==6.0.1
pytest==8.0.0
pytest-cov==4.1.0
schedule==1.2.1
